Updates to the tracking code that differentiates data type
Updates to the tracking code that differentiates data type between POLARRIS, NEXRAD, and NUWRF-EPIC. This is a parameter passed at input.

Additionally, merge/split and unified dataset code has been removed as both are now merged in tobac 1.4.
This has been updated in the environment.yml file as well as a requirement.
However, the additional package xrf (https://github.com/xarray-contrib/xwrf) which is needed to neatly read-in the NUWRF data (currently) requires a higher version of Python (>3.7), and is commented out in the environment.yml file until this is resolved.