Updated tracking script to add user defined variables
This update allows the user to input as a set of strings with no spaces in the command line, additional variables to pull from the files. The max at each feature is found and reported in the Features.nc file.
Example usage is for type="NUWRF" --extra_vars="QNICE","QICE"
This would pull out two additional variables from the NUWRF file, QNICE and QICE. The maximum of each of these is reported in the Features as max_(variable name). Additional documentation in the script.